Added tests for buildDeployment and moved buildSPC tests to kv_util_test
