+++ pkg/controller/keyvault/ingress_secret_provider_class_test.go
@@ -6,7 +6,6 @@ package keyvault
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"testing"
 
 	"github.com/go-logr/logr"
@@ -27,7 +26,6 @@ import (
 	"github.com/Azure/aks-app-routing-operator/pkg/controller/testutils"
 	"github.com/Azure/aks-app-routing-operator/pkg/manifests"
 	"github.com/Azure/aks-app-routing-operator/pkg/util"
-	kvcsi "github.com/Azure/secrets-store-csi-driver-provider-azure/pkg/provider/types"
 )
 
 var (
@@ -314,94 +312,6 @@ func TestIngressSecretProviderClassReconcilerInvalidURL(t *testing.T) {
 
 	assert.Greater(t, afterErrCount, beforeErrCount)
 	assert.Greater(t, afterRequestCount, beforeRequestCount)
-}
-
-func TestIngressSecretProviderClassReconcilerBuildSPCInvalidURLs(t *testing.T) {
-	invalidURLIng := &netv1.Ingress{
-		Spec: netv1.IngressSpec{
-			IngressClassName: &spcTestIngressClassName,
-		},
-	}
-
-	t.Run("nil annotations", func(t *testing.T) {
-		ing := invalidURLIng.DeepCopy()
-
-		ok, err := buildSPC(ing, &secv1.SecretProviderClass{}, spcTestDefaultConf)
-		assert.False(t, ok)
-		require.NoError(t, err)
-	})
-
-	t.Run("empty url", func(t *testing.T) {
-		ing := invalidURLIng.DeepCopy()
-		ing.Annotations = map[string]string{"kubernetes.azure.com/tls-cert-keyvault-uri": ""}
-
-		ok, err := buildSPC(ing, &secv1.SecretProviderClass{}, spcTestDefaultConf)
-		assert.False(t, ok)
-		require.NoError(t, err)
-	})
-
-	t.Run("url with control character", func(t *testing.T) {
-		ing := invalidURLIng.DeepCopy()
-		cc := string([]byte{0x7f})
-		ing.Annotations = map[string]string{"kubernetes.azure.com/tls-cert-keyvault-uri": cc}
-
-		ok, err := buildSPC(ing, &secv1.SecretProviderClass{}, spcTestDefaultConf)
-		assert.False(t, ok)
-		_, expectedErr := url.Parse(cc) // the exact error depends on operating system
-		require.EqualError(t, err, fmt.Sprintf("%s", expectedErr))
-	})
-
-	t.Run("url with one path segment", func(t *testing.T) {
-		ing := invalidURLIng.DeepCopy()
-		ing.Annotations = map[string]string{"kubernetes.azure.com/tls-cert-keyvault-uri": "http://test.com/foo"}
-
-		ok, err := buildSPC(ing, &secv1.SecretProviderClass{}, spcTestDefaultConf)
-		assert.False(t, ok)
-		require.EqualError(t, err, "invalid secret uri: http://test.com/foo")
-	})
-}
-
-func TestIngressSecretProviderClassReconcilerbuildSPCCloud(t *testing.T) {
-	cases := []struct {
-		name, configCloud, spcCloud string
-		expected                    bool
-	}{
-		{
-			name:        "empty config cloud",
-			configCloud: "",
-			expected:    false,
-		},
-		{
-			name:        "public cloud",
-			configCloud: "AzurePublicCloud",
-			spcCloud:    "AzurePublicCloud",
-			expected:    true,
-		},
-		{
-			name:        "sov cloud",
-			configCloud: "AzureUSGovernmentCloud",
-			spcCloud:    "AzureUSGovernmentCloud",
-			expected:    true,
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			ing := spcTestIngress.DeepCopy()
-			ing.Annotations = map[string]string{
-				"kubernetes.azure.com/tls-cert-keyvault-uri": "https://test.vault.azure.net/secrets/test-secret",
-			}
-
-			spc := &secv1.SecretProviderClass{}
-			ok, err := buildSPC(ing, spc, buildTestSpcConfig("test-msi", "test-tenant", c.configCloud))
-			require.NoError(t, err, "building SPC should not error")
-			require.True(t, ok, "SPC should be built")
-
-			spcCloud, ok := spc.Spec.Parameters[kvcsi.CloudNameParameter]
-			require.Equal(t, c.expected, ok, "SPC cloud annotation unexpected")
-			require.Equal(t, c.spcCloud, spcCloud, "SPC cloud annotation doesn't match")
-		})
-	}
 }
 
 func buildTestSpcConfig(msiClient, tenantID, cloud string) *config.Config {
